Import memchr implementation from cortex-strings for Cortex A15
Benchmarks on a Nexus 10 have shown the cortex-strings implementation
of memchr to be 60%-70% faster (depending on input)

Change-Id: I525587d8663c881e9059b26b19409b0c563b2a18
Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>

Conflicts:
	libc/arch-arm/generic/generic.mk
	libc/arch-arm/krait/krait.mk

libc: arm: Slight improvement to memchr.

When this memchr was submitted to newlib years back, ldmia was changed to ldrd.
It was suggested that it would be faster on A15, without a hit to A8/A9;
in fact, my A9 testcase showed a measurable improvement with this change.